test(folder-handle): repair stale folder-handle-states e2e spec
The spec was broken on three independent counts, all pre-existing:
- Stale import (`graph/folder-test-helpers`); the helper moved to
  `graph/folder/` in 625941e42, so the file failed to even load.
- State 3 asserted the chevron was a cytoscape `background-image`, but it
  has been a FolderHandleService DOM-overlay chip since the atlas-blur fix.
- 533 lines, over the 500-line per-file limit.

Repair:
- Fix the import path.
- Rewrite State 3 to assert the DOM chip + chevron glyph for the auth folder
  and clip-shot the chip's real rect (no bbox coordinate math).
- Drive collapse (State 6) and re-expand (State 8) through the shared
  `clickFolderChevron` helper — a real click + hit-test on the actual chevron
  button — instead of fragile body-bbox coordinates and the removed dbltap.
- Extract the fixture + screen helpers into a co-located *.fixtures.ts
  (spec 363 lines, fixtures 147 — both under the limit).
